requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dblpy==0.4.0
decorator==4.4.2
discord.py==1.4.1
idna==2.10
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
multidict==4.7.6
mypy-extensions==0.4.3
mysqlclient==2.0.1
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.0
ratelimiter==1.2.0.post0
redis==3.5.3
regex==2020.7.14
requests==2.24.0
SQLAlchemy==1.3.19
toml==0.10.1
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
yarl==1.5.1
